requirements.txt
Django==1.10
gunicorn==19.6.0
Pillow==3.4.0
psycopg2==2.6.2
whitenoise==3.2.1
